edge_ros2/requirements.txt
# ─── Footwatch Edge — Python Dependencies ────────────────────────────────────
# Install: pip install -r requirements.txt
# Works on both Ubuntu (x86_64) and Raspberry Pi 400 (arm64)

# Computer vision
opencv-python-headless>=4.8.0

# Detection + tracking
ultralytics>=8.2.0
onnxruntime>=1.17.0             # CPU-only; use onnxruntime-gpu on Jetson

# OCR
paddlepaddle>=2.6.0
paddleocr>=2.7.0

# MQTT
paho-mqtt>=1.6.1

# System metrics
psutil>=5.9.0

# Utilities
numpy>=1.24.0
requests>=2.31.0

# Kalman filtering (optional — enhanced tracking)
filterpy>=1.4.5

# Prometheus metrics (optional — health node)
prometheus-client>=0.20.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-timeout>=2.3.0
jsonschema>=4.21.0
